new JS file imported as a module
